Restructure project into Python package
