+++ attentionhtr/test.py
@@ -13,9 +13,9 @@ import torch.nn.functional as F
 import numpy as np
 from nltk.metrics.distance import edit_distance
 
-from utils import CTCLabelConverter, AttnLabelConverter, Averager
-from dataset import hierarchical_dataset, AlignCollate
-from model import Model
+from .utils import CTCLabelConverter, AttnLabelConverter, Averager
+from .dataset import hierarchical_dataset, AlignCollate
+from .model import Model
 device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
